ci: Switch to GitHub Container Registry
This commit updates the CI workflows to pull the CI image from the
GitHub Container Registry (GHCR) instead of the DockerHub, as part of
the effort to move away from using the DockerHub.

The rationale is as follows:

1. DockerHub imposes an impractical pull rate limit (200 pulls per
   6-hour period) that can easily be exceeded in a high volume CI like
   that of the Zephyr Project.

2. GitHub Container Registry is provided by the GitHub and offers
   better integration with the GitHub Actions (e.g. access permissions
   and tokens).

3. DockerHub offers very limited features for free accounts compared to
   the GitHub Container Registry.

Note that the CI image is still available on both GitHub Container
Registry and DockerHub, and third-party users are free to choose
whichever they prefer.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 43a56c2a2d8bd6b75543a7dd2e36e00c30ab0c4d
Change-Id: I3778ca429b97b67f7ce007e02a7afed2f364a1ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3596483
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>