samples, tests: Add z_libc_partition to all test domains
When a memory domain is initialized, the z_libc_partition must be
included so that critical libc-related data can be accessed.

On ARM processors without TPIDRURO when THREAD_LOCAL_STORAGE is enabled,
this includes the TLS base pointer, which is used for several
thread-local variables in the kernel.

Signed-off-by: Keith Packard <keithp@keithp.com>
GitOrigin-RevId: 6f9f8c1e32351c8631b92b3ee59b075bf9880fa1
Change-Id: Id0e87f1a0222451e5040d1c0ba4a68494dea26dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3613436
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>