dts: stm32wb55Xg: fix sram size
STM32WB55xG MCUs include 256 KiB of SRAM split into three banks.
The size of the main bank is 192 KiB, and not 96 KiB as it was
specified in the device tree. This commit fixes the issue and
also updates the definition of the NUCLEO-WB55 board, based on
a STM32WB55RG MCU.

Signed-off-by: Thomas Altenbach <taltenbach@witekio.com>
GitOrigin-RevId: 61e250d511a95e1fc5e69dd09a839ba693500cf5
Change-Id: I7437ee36c3731cb821a5711dda510baeca55bd0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3608293
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>