I2S_MCUX: Fixup I2S MCUX Audio PLL Rate Calculation and Reg Writes
This PR Fixes the Audio PLL Rate Calculation (there was an additional
divide / 8 which is not necessary and does not appear in similar
calculations in example code from the SDK).

Additionally, it adjusts the SAI .dtsi to more correctly configure the
mclk rate, and adds comments specifying what the regististers mean.

Signed-off-by: Nickolas Lapp <nickolaslapp@gmail.com>
GitOrigin-RevId: 521e093cde262d892d8e6e53b915619e9e706fa0
Change-Id: Ib06a3628f027819d1d3a24c1edaccaf49dc8faf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3657370
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>