twister: allow filtering based on modules
Support new keyword in the testcase.yaml to allow filtering based on
enabled modules. This is useful when you only whitelist a limited set of
modules in your project and want to only run those samples that have
their dependencies satisfied and those modules available.

The option allows adding multiple modules, all of which need to be
enabled in the manifest.

Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: f429f339fc8fb9adcb5d87559cabe0f164106601
Change-Id: I50d06846cf8e77039f9c5ae271f981bebd8fce47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3596488
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>