drivers: xen: keep track of missed events on event channels
The current implementation for events channel is using an empty
callback for every unbind channel and the interrupt is clearing
every event and calling the callback.
However in a scenario where a domain fires a notification when
another has not yet bind the channel, the event will be missed.

To address this limitation, this commit is keeping track of
missed event channel notification when the empty callback is
used, a function to retrieve and clear the missed event is
introduced.

(cherry picked from commit 58efbc6f4f4e2a8f67f636166f097f9f7860b0bc)

Original-Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
GitOrigin-RevId: 58efbc6f4f4e2a8f67f636166f097f9f7860b0bc
Change-Id: I8230f584dea9e33ec9e4445b0e04452f26bb8b32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3707143
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>