ipc_service: static_vrings: Set WQ default type to PRIO_COOP
In 92d8329d5b6 a new DT property was introduced to set the WQ priority
of the instance. The fallback value when the property was not present
was arbitrarily set to <0 PRIO_PREEMPT>.

The problem is that this value is actually changing the behaviour for
the code that is not explicitly setting the DT property, breaking in
some cases the existing code.

Move the default value to <0 PRIO_COOP> to give the old code a
consistent behaviour before and after the 92d8329d5b6 commit.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>
GitOrigin-RevId: 7f51907fda2dd30f18d01262f06c175714a9be3f
Change-Id: If82c6d262a3f202b68337de2a05de2de771db6dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3613449
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>