boards: arduino_nano_33_ble: Correct polarity of pwm1 channel 0
This is a follow-up to commit f4a0ddd8afa1c2b99a30dcd940f34304ce678c93.

According to the schematic, the LED connected to the P1.09 pin is
active high. Therefore, the PWM1 instance that is configured to drive
the LED should not use the "nordic,invert" property.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: a4fa8c6ac01383dec092648d76e97bb09b80f8d3
Change-Id: I342301d2cc4424f2f1fe2bfba59b1011e3260937
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3602456
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>