arch/x86: Use EFI console as default printk handler
Where we have access to a bootstrap UEFI environment, it's productive
to use that console as the default printk handler.  That avoids the
bringup hassle of trying to configure UART settings blindly, as has
been customary.  It also emits nice text to the framebuffer on devices
with no serial port or other debug harness at all.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: e931b7ba4707cfb89cda1e89dcfed50563d79bc8
Change-Id: Iacac1201158abf89b5b21fefeae6c0559873fd64
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3627732
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>