boards: arm: legend: fix PWM LEDs period
The period was 255 nsec, a value that doesn't make much sense when
driving an LED. Since the period cell is rarely used nowadays, the
value was probably copy&pasted or a random value was added since it is
required. A period of 20 msec has been chosen as most other boards do
for PWM LEDs.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: fe41c54bcd189d7358701b4d552f7cd2fd2605cb
Change-Id: Ie3b774c1bda35bae6a79057df735d59995f221a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3605348
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>