boards: arm: twr_ke18f: fix PWM LEDs period
The period was set to 60 usec, a value that doesn't make much sense in
the context of PWM driven LEDs. Since the period cell was rarely used,
the value was likely added because it is required, but was never used. A
period of 20 msec has been chosen as most other boards do.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: ab71e6bdf0fe1558c038348a48c0a1bc88b41d19
Change-Id: I5f70293fb4eba8f546be17dac16244459a0064ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3598994
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>