doc: Fix the layout of POSIX-related tables
Several tables in the section on POSIX support have a column
"Supported" which is supposed to indicate whether a feature,
a function, or similar, is supported in Zephyr or not. This
indication is done by placing character "+" in relevant cells.

However, in restructuredText, "+" is a sign for an element of
a bulleted list (as is the case with "*" and several other
characters). This, in turn, causes very inefficient rendering of
all these cells with "+", since restructuredText adds spacing
before and after any bulleted list.

Fix this by replacing all "+" characters by a simple "yes".

Signed-off-by: Aleksandar Markovic <aleksandar.markovic.sa@gmail.com>
GitOrigin-RevId: a07774eadabbb7eea5ebd605cb23287b6c8b2c0e
Change-Id: I2ac8bcede3922e5d547096385aad8c1c00c05a09
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3608292
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>