semihosting: fix inline assembly output dependency
Commit d8f186aa4a18 ("arch: common: semihost: add semihosting
operations") encapsulated semihosting invocation in a per-arch
semihost_exec() function. There is a fixed register variable declaration
for the return value but this variable is not listed as an output
operand to respective inline assembly segments which is an error.
This is not reported as such by gcc and the generated code is still OK
in those particular instances but this is not guaranteed, and clang
does complain about such cases.

Signed-off-by: Nicolas Pitre <npitre@baylibre.com>
GitOrigin-RevId: f61b8b8c1611ce4065a54cc474712fed27abb003
Change-Id: I04bf2d040067657847d13f07a3833132695cb9d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3604596
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>