dts: bindings: Add memory-region property in bindings
Since the introduction of the `zephyr,memory-region` compatible
to create new memory region, drivers can now use the newly defined
region in the DT by using the macro LINKER_DT_NODE_REGION_NAME()
on the node_id to retrieve the region name and possibly allocated data
or variables in that region (or for whatever other use).
This is assuming though that the driver knows in advance the node_id to
reference.

A better approach is to add a new generic 'memory-region' property
that can be used by any driver to reference a region by using a phandle,
so that the driver can use that to retrieve the memory-region of interest.

(cherry picked from commit a471be3a41a3361bf835d3f260d618b13d03fe24)

Original-Signed-off-by: Adam Wojasinski <adam.wojasinski@nordicsemi.no>
Original-Signed-off-by: Carlo Caione <ccaione@baylibre.com>
GitOrigin-RevId: a471be3a41a3361bf835d3f260d618b13d03fe24
Change-Id: I9fdda4e3ab995487d2da09ac61fdad309467c147
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3691485
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>