boards: imx8mp: partial pin control support
Add partial pin control support for the imx8mp. Since the UART driver is
not currently enabled, pin control cannot be tested on this platform.
Therefore, only the DTS definitions required to set the pinmux options
are present for this platform, and are not being applied (since
CONFIG_PINCTRL=n).

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: b6377cce6a56bb3849e2a8c9c66cd8623bbb28b0
Change-Id: Iebd0b1a41a1afe4b17434c8102b921feafdf28e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3647598
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>