MAINTAINERS: Add Toolchain Integration entry
This commit adds the MAINTAINERS entry for the Toolchain Integration,
with @tejlmand as the maintainer and @stephanosio as a collaborator.

The purpose of this is to provide a corresponding MAINTAINERS entry for
the "area: Toolchains" and differentiate the toolchain integration-
related maintenance tasks from the general build system maintenance
tasks.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: fbe8a673d4c3aa5627d6e2a2a2e9da6cc3341f50
Change-Id: Ibcc8475d3c52e50e63a01cafc6ec7f730db36573
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3614105
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>