editorconfig: Add C++ indentation style
This commit adds the indentation style for the C++ source and header
files.

Note that the C++ indentation style has been configured to match that
of the C files because the C++ source files we currently have in the
Zephyr repository follow the style we enforce for the C files.

In the future, it may be preferable to follow the indentation style
recommended by one of the common C++ style guides, such as the Google
C++ Style Guide which recommends 2 spaces for indentation.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 3a7cf87821176bee781d514affdaa71fe7ac529f
Change-Id: I0346b6ef1674f838c6066c0c755709c797ff0583
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3614843
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>