Bluetooth: Audio: MCC - serialized subscription after discovery
The media control client used to subscribe to GMCS characteristics as
the characteristics were discovered, in effect queuing ATT requests.

Since then, the ATT MTU exchange has become automatic, leading to
larger ATT MTUs and more discovered characteristics per ATT response,
which again lead to the client enqueuing more subscriptions and
running out of buffers, with discovery failing as a result.

This commit moves the subscription to GMCS characteristics to a
separate sequence, so that it happens serially after discovery, rather
than in parallel with the discovery.

- Remove separate discovery parameters for the characteristics - they
  are no longer needed.
- Add a subscription callback and a couple of subscription functions

The subscription for characteristics for the included OTS has not been
modified.  Only two of these are subscribed to, and doing that is
currently not a problem.  And I need to better understand the
interaction with OTS and how these subscriptipons are used.

Refactor subscription to get chaining into callback

(cherry picked from commit 0142bf94845660af9e33d5c2374c02b410de0a45)

Original-Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>
GitOrigin-RevId: 0142bf94845660af9e33d5c2374c02b410de0a45
Change-Id: I95768415572d02eecbc8ba9a92a7f744e0389291
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3753921
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>