samples: drivers: stm32u5x octospi enables DMA
Adds the dma transfer for the octoSPI on NOR octo Flash
of the b_u585i_iot02a disco kit.
The channel for the GPDMA is 0-15 (0-7 for DMA1, 8-15 for DMA2)
The channel 12 to 15 are used for GPDMA transfers
to/from external memories.
The request is 41 for the OCTOSPI2.

(cherry picked from commit 055c1c637e5d62b53a237c2ce34f3867ae373c20)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: 055c1c637e5d62b53a237c2ce34f3867ae373c20
Change-Id: Ie35d2a12e688fb93502d21919eb0c8cebfb98383
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4048039
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>