driver: can: stm32fd: add stm32u5 clock selection
The STM32u% series of processors has a unique set of clock sources for
the FDCAN peripheral. This brings the selection in line with the
existing can_stm32fd clock selection Kconfigs.

This change was tested on a proprietary board using the STM32U5 series
which exposes the CAN pins of the SOC using a transciever on a live CAN
bus as well as on the nucleo_g474re board from ST in loopback mode.

HSE and PLL1Q tests run and all passed.

PLL2P is not currently supported by the clock drivers for STM32U5, and
as such is currently untested. When this support is added, the driver
should be able to use this clock without issue.

When changes from #42097 are merged this fix should be deprecated in
favor of using the methods outlined there.

Signed-off-by: Peter Maxwell Warasila <madmaxwell@soundcomesout.com>
GitOrigin-RevId: c91a70e130ea5c1591c126675dc19890eaa21384
Change-Id: Ia44b891e92a909b8ae4e25a6831b991bb7108f99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3631440
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>