i2c: Adds i2c_transfer_async API for I2C
i2c_transfer_async is the asynchronous version of i2c_transfer where
the completion of the transfer is notified to the caller using a callback.

This can be used in conjuction with the common callbacks and datatypes
in async.h for directly doing an async transfer with an IPC object
to notify a thread.

(cherry picked from commit a45ece6984d0ef576f6a2bdd604b1842a0872f73)

Original-Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: a45ece6984d0ef576f6a2bdd604b1842a0872f73
Change-Id: Iee30f548ce254e6da71d85483d562c3ebf4a2d23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3829380
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>