kernel: sched: Change cpu pin only for not executing threads
Do not allow changing the CPU which a thread is pinned when it is
already being executed. This allows further optimizations in some
platforms with incoherent memory since we can safely assume that the
thread will run in the same CPU and avoid invalidate / flush the
cache during context switches.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: 551038e74865567f453beecbbf9ea4c9c03663cc
Change-Id: I254b031db99de0bf829d1367e9cf55ec2ff9de99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3628003
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>