drivers/sensor: lsm6dso: Remove LE to CPU conversions
In ST HAL v2.00, the functions to get the raw sensors values, e.g.
lsm6dso_acceleration_raw_get(), convert from little-endian to CPU.

Previous versions of ST HAL didn't do this.

The conversion here in the driver is converting a second time.  It's not
an issue on a LE system, the conversion is a no-op, but on a BE system
it would be broken.

(cherry picked from commit c269692b1d6d562dcde3c3ad422e75af4d5823ad)

Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
GitOrigin-RevId: c269692b1d6d562dcde3c3ad422e75af4d5823ad
Change-Id: I755aa64a19adda94decb7c23e3150e1d2728e976
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3670855
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>