drivers/sensor: lsm6dso: Add support for lsm6dso32
This sensor is virtually identical to the lsm6dso.  The only difference
is the accelerometer ranges are double those of the lsm6dso.

Use the same driver.  The difference is detected by using "st,lsm6dso32"
as the first compatible entry, followed by "st,lsm6dso".

An bit flag in the existing accel_range config field is used to check if
the chip is the doubled range or not.

(cherry picked from commit 094345629b15b007ef1e3927769c396083870302)

Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
GitOrigin-RevId: 094345629b15b007ef1e3927769c396083870302
Change-Id: I52244e4cd0aa3d82a49576ab6eb919dac5d86914
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3670857
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>